move max connection idle manager to core
